+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+- Add ODT export functionality for assessment tables
+- Add checkbox in procedure settings to expand procedure description in public view on page load
+- Use external Geocoder API as service for address auto-suggestions
 
 ## v4.11.0 (2025-08-27)
 - Allow to edit custom field type singleSelect